Fb.py fixed
Scraping actual Facebook site is a little problematic as getting its relative xpath is difficult. So we are using (https://mbasic.facebook.com/) which has simple HTML format and loads up quickly. Only thing is that after login the URL changes, so redirecting it again to previous URL is required or else it throws element not found error. 
It's now complete